fix bugs

+++ src/Vitrina.UseCases/ProjectPage/DeleteProjectPage/DeleteProjectPageCommandHandler.cs
@@ -1,4 +1,5 @@
 using MediatR;
+using Saritasa.Tools.Domain.Exceptions;
 using Vitrina.Infrastructure.Abstractions.Interfaces.Repositories;
 
 namespace Vitrina.UseCases.ProjectPage.DeleteProjectPage;
@@ -12,6 +13,10 @@ public class DeleteProjectPageCommandHandler(IProjectPageRepository repository)
     {
         var page = await repository.GetByIdAsync(request.Id, cancellationToken);
         page.ThrowExceptionIfNoAccessRights(request.IdAuthorizedUser);
+        if (page.ProjectId != null)
+        {
+            throw new DomainException("The project page cannot be deleted while the associated project exists.");
+        }
         await repository.Delete(request.Id, cancellationToken);
         await repository.SaveChangesAsync(cancellationToken);
     }
